Add a pictures endpoint.
